.developing rainbow
git-svn-id: svn://ultimatepp.org/upp/trunk@3582 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ rainbow/LinuxFb/Proc.cpp
@@ -143,7 +143,7 @@ void FBInit(const String& fbdevice)
 	// Map the device to memory
 	fbp = (char *)mmap(0, screensize, PROT_READ | PROT_WRITE, MAP_SHARED,
 					fbfd, 0);
-	if ((int)fbp == -1) {
+	if ((intptr_t)fbp == -1) {
 		printf("Error: failed to map framebuffer device to memory.\n");
 		exit(4);
 	}
